+++ src/app/screens/Auth/EmailLoginScreen.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { StyleSheet, Text, Image, View, TouchableHighlight, SafeAreaView, Button, Alert, Platform, StatusBar, Dimensions,ImageBackground } from 'react-native'
-import Input from '../../components/Input';
-import LoginButton from '../../components/LoginButton';
-import COLORS from '../../constants';
+import Input from '../../../components/Input';
+import LoginButton from '../../../components/LoginButton';
+import { COLORS } from '../../../constants';
 import { FontAwesome5 } from '@expo/vector-icons';
-import PasswordInput from '../../components/PasswordInput';
-import Data from "../../Data/data.json"
+import PasswordInput from '../../../components/PasswordInput';
+import Data from "../../../Data/data.json"
 import { useState } from 'react';
 import { Link } from 'expo-router'
 
@@ -26,7 +26,7 @@ function EmailLoginScreen() {
         <SafeAreaView style={styles.container}>
         <View style={styles.parent}>
             <View style={styles.upper}>
-                <ImageBackground source={require("../../../assets/headImage.png")} style={styles.Image}>
+                <ImageBackground source={require("../../../../assets/headImage.png")} style={styles.Image}>
                     <View style={styles.IntroView}>
                         <Text style={styles.IntroText}>Glad to see you!!</Text>
                     </View>
