+++ src/app/screens/Home/DetailCategoryScreen.tsx
