+++ src/components/Paginator.tsx
