+++ src/components/PasswordInput.tsx
@@ -1,6 +1,6 @@
 import React,{useState} from 'react';
 import { StyleSheet, Text, Image, View, TouchableHighlight, SafeAreaView, Button, Alert, Platform, StatusBar, Dimensions,TextInput} from 'react-native'
-import COLORS from '../constants';
+import { COLORS } from '../constants';
 import { Ionicons } from '@expo/vector-icons';
 import { Feather } from '@expo/vector-icons';
