+++ src/constants/Icons.ts
@@ -1,0 +1,41 @@
+import messageIcon from "../../assets/icons/Message.png"
+import callIcon from "../../assets/icons/Call.png"
+import  filterIcon from "../../assets/icons/Filter.png"
+import searchIcon from "../../assets/icons/Homesearch.png"
+import locationIcon from "../../assets/icons/Location.png"
+import notificationIcon from "../../assets/icons/Notification.png"
+import timeIcon from "../../assets/icons/Time Circle.png" 
+import logoutIcon from "../../assets/icons/LogoutIcon.png" 
+import profileIcon from "../../assets/icons/ProfileIcon.png"
+import discountIcon from "../../assets/icons/discountIcon.png" 
+import documentIcon from "../../assets/icons/DocumentIcon.png"
+import  questionIcon from "../../assets/icons/bi_question-circleIcon.png"
+import settingsIcon from "../../assets/icons/SettingIcon.png"
+import headphoneIcon from "../../assets/icons/headphoneIcon.png" 
+import rightIcon from "../../assets/icons/toTheRigthIcon.png" 
+import filterWithBackground from "../../assets/icons/filtterWithBackground.png" 
+import leftArrowWithBackground from "../../assets/icons/leftArrowWithBackground.png" 
+import currentLocationIcon from "../../assets/icons/currentLocationIcon.png" 
+
+
+
+export {
+    currentLocationIcon,
+    profileIcon,
+    discountIcon,
+    documentIcon,
+    questionIcon,
+    settingsIcon,
+    headphoneIcon,
+    messageIcon,
+    callIcon,
+    filterIcon,
+    searchIcon,
+    locationIcon,
+    notificationIcon,
+    timeIcon,
+    logoutIcon,
+    rightIcon,
+    filterWithBackground,
+    leftArrowWithBackground 
+}
