+++ src/constants/Images.ts
@@ -1,0 +1,23 @@
+import welcomeImage from "../../assets/images/welcomeImage.png";
+import HollaImage from "../../assets/images/hollaImage.png";
+import findParkingImage from "../../assets/images/findParkingImage.png";
+import mall from "../../assets/images/grahaMall.png";
+import HomebackgroundImage from "../../assets/images/maskBackgroundImage.png";
+import dianaImage from "../../assets/images/dianaImage.png";
+import exploreImage from "../../assets/images/explore.png" 
+import locationImage from "../../assets/images/locationBtn.png"
+import parkDetailImage from "../../assets/images/parkDetail.png"
+
+
+
+export { 
+    welcomeImage, 
+    HollaImage, 
+    findParkingImage, 
+    mall,
+    HomebackgroundImage,
+    locationImage, 
+    dianaImage, 
+    exploreImage, 
+    parkDetailImage
+};
